refactor: Move validator formatting to `validator.rs`
