+++ jsonschema/src/keywords/any_of.rs
@@ -1,9 +1,8 @@
 use crate::{
     compilation::{compile_validators, context::CompilationContext, JSONSchema},
     error::{error, no_error, ErrorIterator, ValidationError},
-    keywords::{format_vec_of_validators, Validators},
     paths::InstancePath,
-    validator::Validate,
+    validator::{format_vec_of_validators, Validate, Validators},
 };
 use serde_json::{Map, Value};
