+++ jsonschema/src/validator.rs
@@ -1,4 +1,6 @@
-use crate::{compilation::JSONSchema, error::ErrorIterator, paths::InstancePath};
+use crate::{
+    compilation::JSONSchema, error::ErrorIterator, keywords::BoxedValidator, paths::InstancePath,
+};
 use serde_json::Value;
 use std::fmt;
 
@@ -19,4 +21,45 @@ impl fmt::Debug for dyn Validate + Send + Sync {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         f.write_str(&self.to_string())
     }
+}
+
+pub(crate) type Validators = Vec<BoxedValidator>;
+
+pub(crate) fn format_validators(validators: &[BoxedValidator]) -> String {
+    match validators.len() {
+        0 => "{}".to_string(),
+        1 => {
+            let name = validators[0].to_string();
+            match name.as_str() {
+                // boolean validators are represented as is, without brackets because if they
+                // occur in a vector, then the schema is not a key/value mapping
+                "true" | "false" => name,
+                _ => format!("{{{}}}", name),
+            }
+        }
+        _ => format!(
+            "{{{}}}",
+            validators
+                .iter()
+                .map(|validator| format!("{:?}", validator))
+                .collect::<Vec<String>>()
+                .join(", ")
+        ),
+    }
+}
+
+pub(crate) fn format_vec_of_validators(validators: &[Validators]) -> String {
+    validators
+        .iter()
+        .map(|v| format_validators(v))
+        .collect::<Vec<String>>()
+        .join(", ")
+}
+
+pub(crate) fn format_key_value_validators(validators: &[(String, Validators)]) -> String {
+    validators
+        .iter()
+        .map(|(name, validators)| format!("{}: {}", name, format_validators(validators)))
+        .collect::<Vec<String>>()
+        .join(", ")
 }
